+++ src/CssSorterPackage.cs
@@ -1,15 +1,55 @@
-﻿using Microsoft.VisualStudio.Shell;
+﻿using Microsoft.VisualStudio;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
 using System;
 using System.Runtime.InteropServices;
+using System.Threading;
+using Tasks = System.Threading.Tasks;
 
 namespace CssSorter
 {
     [PackageRegistration(UseManagedResourcesOnly = true)]
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
     [Guid(PackageGuids.guidPackageString)]
+    [ProvideOptionPage(typeof(Options), "Web", Vsix.Name, 101, 111, true, new[] { "css", "sort" }, ProvidesLocalizedCategoryName = false)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
     public sealed class CssSorterPackage : AsyncPackage
     {
+        private static Options _options;
+        private static object _syncRoot = new object();
 
+        public static Options Options
+        {
+            get
+            {
+                if (_options == null)
+                {
+                    lock (_syncRoot)
+                    {
+                        if (_options == null)
+                        {
+                            LoadPackage();
+                        }
+                    }
+                }
+
+                return _options;
+            }
+        }
+
+        protected override Tasks.Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
+        {
+            _options = (Options)GetDialogPage(typeof(Options));
+
+            return base.InitializeAsync(cancellationToken, progress);
+        }
+
+        private static void LoadPackage()
+        {
+            var shell = (IVsShell)GetGlobalService(typeof(SVsShell));
+
+            if (shell.IsPackageLoaded(ref PackageGuids.guidPackage, out IVsPackage package) != VSConstants.S_OK)
+                ErrorHandler.Succeeded(shell.LoadPackage(ref PackageGuids.guidPackage, out package));
+        }
     }
 }